+++ SwitchStartupProjectShared/Extensions.cs
@@ -1,0 +1,10 @@
+// Extensions.cs
+using System;
+
+namespace SwitchStartupProject
+{
+    static class ExtensionsList
+    {
+        public const string extensionSharedProject = ".shproj";
+    };
+}